requirements.txt
numpy>=1.24.0
matplotlib>=3.6.0
plotly>=5.15.0
dash>=2.12.0
scipy>=1.10.0
numba>=0.57.0
opencv-python>=4.8.0
pillow>=9.5.0
pyopengl>=3.1.0
moderngl>=5.8.0
pygame>=2.5.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
seaborn>=0.12.0
pandas>=1.5.0